+++ bin/make/frameworks.sh
@@ -1,16 +1,11 @@
 #!/usr/bin/env bash
 
 set +e
-
-# Force Xcode 8 CoreSimulator env to be loaded so xcodebuild does not fail.
-for try in {1..4}; do
-  xcrun simctl help &>/dev/null
-  sleep 1.0
-done
-
+source bin/simctl.sh
+ensure_valid_core_sim_service
 set -e
 
-source bin/log_functions.sh
+source bin/log.sh
 
 if [ -z "${FBSIMCONTROL_PATH}" ]; then
   if [ -e "../FBSimulatorControl" ]; then