+++ iOSDeviceManager/Devices/Device.h
@@ -7,32 +7,6 @@
 #import "iOSReturnStatusCode.h"
 #import "CodesignIdentity.h"
 
-@protocol DVTApplication;
-
-@interface FBiOSDeviceOperator (iOSDeviceManagerAdditions)
-
-- (void)fetchApplications;
-- (BOOL)killProcessWithID:(NSInteger)processID error:(NSError **)error;
-
-// The keys-value pairs that are available in the plist returned by
-// #installedApplicationWithBundleIdentifier:error:
-+ (NSDictionary *)applicationReturnAttributesDictionary;
-- (NSDictionary *)AMDinstalledApplicationWithBundleIdentifier:(NSString *)bundleID;
-
-// These will probably be moved to FBDeviceApplicationCommands
-- (BOOL)isApplicationInstalledWithBundleID:(NSString *)bundleID error:(NSError **)error;
-- (BOOL)launchApplication:(FBApplicationLaunchConfiguration *)configuration
-                    error:(NSError **)error;
-
-// Originally, we used DVT APIs to install provisioning profiles.
-// Facebook is migrating from DVT to MobileDevice (Apple MD) APIs.
-// If we find there is a problem with the MobileDevice API we can
-// fall back on the DVT implementation.
-// - (BOOL)DVTinstallProvisioningProfileAtPath:(NSString *)path error:(NSError **)error;
-- (BOOL)AMDinstallProvisioningProfileAtPath:(NSString *)path error:(NSError **)error;
-
-@end
-
 @interface FBXCTestRunStrategy (iOSDeviceManagerAdditions)
 
 /**
@@ -81,7 +55,6 @@
 + (iOSReturnStatusCode)generateXCAppDataBundleAtPath:(NSString *)path
                                            overwrite:(BOOL)overwrite;
 
-- (FBiOSDeviceOperator *)fbDeviceOperator;
 - (iOSReturnStatusCode)launch;
 - (iOSReturnStatusCode)kill;
 
@@ -109,6 +82,9 @@
 
 //TODO: this should accept Env and Args
 - (iOSReturnStatusCode)launchApp:(NSString *)bundleID;
+
+- (BOOL)launchApplicationWithConfiguration:(FBApplicationLaunchConfiguration *)configuration
+                                     error:(NSError **)error;
 - (iOSReturnStatusCode)killApp:(NSString *)bundleID;
 - (BOOL)shouldUpdateApp:(Application *)app statusCode:(iOSReturnStatusCode *)sc;
 - (iOSReturnStatusCode)isInstalled:(NSString *)bundleID;